DEV-3150: change 'pretend' to 'dry-run'
git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@13409 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ code/base/tim-get-tool/src/org/terracotta/modules/tool/commands/UpgradeCommand.java
@@ -26,7 +26,7 @@ public class UpgradeCommand extends ModuleOperatorCommand {
 
   private static final String             LONGOPT_OVERWRITE = "overwrite";
   private static final String             LONGOPT_FORCE     = "force";
-  private static final String             LONGOPT_PRETEND   = "pretend";
+  private static final String             LONGOPT_DRYRUN   = "dry-run";
   private static final String             LONGOPT_NOVERIFY  = "no-verify";
 
   private final Collection<InstallOption> installOptions;
@@ -34,7 +34,7 @@ public class UpgradeCommand extends ModuleOperatorCommand {
   public UpgradeCommand() {
     options.addOption(buildOption(LONGOPT_OVERWRITE, "Install anyway, even if already installed"));
     options.addOption(buildOption(LONGOPT_FORCE, "Synonym to overwrite"));
-    options.addOption(buildOption(LONGOPT_PRETEND, "Do not perform actual installation"));
+    options.addOption(buildOption(LONGOPT_DRYRUN, "Do not perform actual installation"));
     options.addOption(buildOption(LONGOPT_NOVERIFY, "Skip checksum verification"));
     arguments.put("file", "The path to tc-config.xml");
     installOptions = new ArrayList<InstallOption>();
@@ -59,7 +59,7 @@ public class UpgradeCommand extends ModuleOperatorCommand {
   public void execute(CommandLine cli) {
     if (cli.hasOption(LONGOPT_FORCE)) installOptions.add(InstallOption.FORCE);
     if (cli.hasOption(LONGOPT_OVERWRITE) || cli.hasOption(LONGOPT_FORCE)) installOptions.add(InstallOption.OVERWRITE);
-    if (cli.hasOption(LONGOPT_PRETEND)) installOptions.add(InstallOption.PRETEND);
+    if (cli.hasOption(LONGOPT_DRYRUN)) installOptions.add(InstallOption.DRYRUN);
     if (cli.hasOption(LONGOPT_NOVERIFY)) installOptions.add(InstallOption.SKIP_VERIFY);
     try {
       process(cli);
@@ -136,7 +136,7 @@ public class UpgradeCommand extends ModuleOperatorCommand {
     }
 
     // save original file to .original if found newer module
-    if (updateConfig) {
+    if (updateConfig && !cli.hasOption(LONGOPT_DRYRUN)) {
       File originalFile = new File(tcConfigPath.getAbsolutePath() + ".original");
       FileUtils.copyFile(tcConfigPath, originalFile);
       out.println("Your original config file has been saved to " + originalFile);
